+++ lib/ace_cmd.rb
@@ -6,3 +6,6 @@ require "ace_cmd/errors"
 require "ace_cmd/result"
 require "ace_cmd/success"
 require "ace_cmd/failure"
+
+require "ace_cmd/ace_command"
+require "ace_cmd/ace_callee"
